Move TraceContext propagation into its own namespace
