+++ adapters/ethon/test/opentelemetry/adapters/ethon/adapter_test.rb
@@ -20,7 +20,7 @@ describe OpenTelemetry::Adapters::Ethon::Adapter do
     # these are currently empty, but this will future proof the test
     @orig_injectors = OpenTelemetry.propagation.http_injectors
     OpenTelemetry.propagation.http_injectors = [
-      OpenTelemetry::Trace::Propagation.http_trace_context_injector
+      OpenTelemetry::Trace::Propagation::TraceContext.http_trace_context_injector
     ]
   end
