+++ adapters/faraday/test/opentelemetry/adapters/faraday/middlewares/tracer_middleware_test.rb
@@ -31,7 +31,7 @@ describe OpenTelemetry::Adapters::Faraday::Middlewares::TracerMiddleware do
     # these are currently empty, but this will future proof the test
     @orig_injectors = OpenTelemetry.propagation.http_injectors
     OpenTelemetry.propagation.http_injectors = [
-      OpenTelemetry::Trace::Propagation.http_trace_context_injector
+      OpenTelemetry::Trace::Propagation::TraceContext.http_trace_context_injector
     ]
   end
