+++ adapters/net_http/test/opentelemetry/adapters/net/http/adapter_test.rb
@@ -23,7 +23,7 @@ describe OpenTelemetry::Adapters::Net::HTTP::Adapter do
     # these are currently empty, but this will future proof the test
     @orig_injectors = OpenTelemetry.propagation.http_injectors
     OpenTelemetry.propagation.http_injectors = [
-      OpenTelemetry::Trace::Propagation.http_trace_context_injector
+      OpenTelemetry::Trace::Propagation::TraceContext.http_trace_context_injector
     ]
   end
