+++ adapters/restclient/test/opentelemetry/adapters/restclient/adapter_test.rb
@@ -22,7 +22,7 @@ describe OpenTelemetry::Adapters::RestClient::Adapter do
     # these are currently empty, but this will future proof the test
     @orig_injectors = OpenTelemetry.propagation.http_injectors
     OpenTelemetry.propagation.http_injectors = [
-      OpenTelemetry::Trace::Propagation.http_trace_context_injector
+      OpenTelemetry::Trace::Propagation::TraceContext.http_trace_context_injector
     ]
   end
