+++ api/test/opentelemetry/trace/propagation/trace_context/trace_parent_test.rb
@@ -5,8 +5,8 @@
 # SPDX-License-Identifier: Apache-2.0
 require 'test_helper'
 
-describe OpenTelemetry::Trace::Propagation::TraceParent do
-  TraceParent = OpenTelemetry::Trace::Propagation::TraceParent
+describe OpenTelemetry::Trace::Propagation::TraceContext::TraceParent do
+  TraceParent = OpenTelemetry::Trace::Propagation::TraceContext::TraceParent
   Trace = OpenTelemetry::Trace
   let(:good) do
     flags = Trace::TraceFlags.from_byte(1)
